Core: lazy docgen middleware
