+++ code/renderers/react/src/docgen/docgen-worker.ts
@@ -13,8 +13,8 @@
  * docgen-relevant TS option (`reactDocgenTypescriptOptions`, including `propFilter`) applies solely
  * to the `react-docgen-typescript` engine used by the legacy manifest path — not here.
  */
-import { STORY_FILE_TEST_REGEXP, getStoryImportPathFromEntry } from 'storybook/internal/common';
-import type { DocgenMiddleware, DocgenProvider } from 'storybook/internal/types';
+import { createLazyDocgenMiddleware } from 'storybook/internal/common';
+import type { DocgenMiddleware } from 'storybook/internal/types';
 
 import { ComponentMetaManager } from '../componentManifest/componentMeta/ComponentMetaManager.ts';
 import { buildDocgenPayload } from './buildDocgen.ts';
@@ -25,41 +25,15 @@ import { buildDocgenPayload } from './buildDocgen.ts';
  * created lazily on the first eligible request and memoized; when TypeScript is unavailable the
  * middleware passes through to the rest of the chain.
  */
-export const createDocgenProvider = (): DocgenMiddleware => {
-  let managerPromise: Promise<ComponentMetaManager | undefined> | undefined;
-
-  const getManager = (): Promise<ComponentMetaManager | undefined> => {
-    if (!managerPromise) {
-      managerPromise = (async () => {
-        try {
-          const ts = await import('typescript');
-          return new ComponentMetaManager(ts);
-        } catch {
-          return undefined;
-        }
-      })();
-    }
-    return managerPromise;
-  };
-
-  return (nextDocgen: DocgenProvider): DocgenProvider =>
-    async (input) => {
-      const storyImportPath = getStoryImportPathFromEntry(input.entry);
-      if (!storyImportPath || !STORY_FILE_TEST_REGEXP.test(storyImportPath)) {
-        return nextDocgen(input);
+export const createDocgenProvider = (): DocgenMiddleware =>
+  createLazyDocgenMiddleware({
+    createManager: async () => {
+      try {
+        const ts = await import('typescript');
+        return new ComponentMetaManager(ts);
+      } catch {
+        return undefined;
       }
-
-      const componentMetaManager = await getManager();
-      if (!componentMetaManager) {
-        return nextDocgen(input);
-      }
-
-      const ours = await buildDocgenPayload(input, { componentMetaManager });
-      if (!ours) {
-        return nextDocgen(input);
-      }
-
-      const downstream = await nextDocgen(input);
-      return { ...downstream, ...ours };
-    };
-};
+    },
+    extract: (componentMetaManager, input) => buildDocgenPayload(input, { componentMetaManager }),
+  });
